Moderate improvements to testing convai2.
