+++ projects/convai2/baselines/language_model/eval_f1.py
@@ -13,7 +13,7 @@ from parlai.core.build_data import download_models
 from projects.convai2.eval_f1 import setup_args, eval_f1
 
 
-def main():
+def main(args=None):
     parser = setup_args()
     parser.set_params(
         model='language_model',
@@ -21,7 +21,7 @@ def main():
         dict_file='models:convai2/language_model/model.dict',
         batchsize=20,
     )
-    opt = parser.parse_args()
+    opt = parser.parse_args(args=args)
     opt['model_type'] = 'language_model'
     fnames = ['model', 'model.dict', 'model.opt']
     download_models(opt, fnames, 'convai2', version='v2.0', use_model_type=True)
